Add footnote for clarity
"You can open the `vscode` folder" is ambiguous as to where the folder
should be opened. Because of this I changed the wording to be a bit
clearer and added a "code ." suggestion as that would be easier for
most developers to follow at the point the previous steps put them into
the vscode folder.