Add warning for path with spaces
When following this guide, the build step was failing with both methods.
I didn't know why it wasn't working until I realized that my path
contained spaces and that was breaking things.

To help future users I added a warning and troubleshooting tip so that
people can avoid this mistake in the future or people who also make
this mistake will have less trouble resolving it.